update Makefile: added clean + format targets
I just added two new targets in the Makefile.
The first one, which is called `clean`, just removes all files that were
generated during a build.
The second one is `format`, and supposes that you have `clang-format`
installed on your computer. When run, it formats all files having a `.c`
or `.h` extension.

I also needed to add `-D_POSIX_SOURCE` on the `CFLAGS` to make the
project build on my machine.